# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/gemini/chat": {
      "post": {
        "description": "接收用户消息，返回 Gemini 模型生成的回复，包含 token 统计信息。",
        "operationId": "ChatController_handleChat",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatMessageParamsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/ChatResponseDto",
                          "description": "Gemini结果",
                          "example": {
                            "content": "你好！很高兴你来这里。有什么我可以帮忙的吗？",
                            "usage": {
                              "promptTokens": 15,
                              "completionTokens": 25,
                              "totalTokens": 40
                            },
                            "model": "gemini-2.5-flash",
                            "responseTime": 1200
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "Gemini 聊天接口",
        "tags": [
          "Gemini 聊天与语音"
        ]
      }
    },
    "/gemini/stream": {
      "post": {
        "description": "使用 VertexAI Gemini 提供流式聊天服务，返回结构化响应数据，包含 token 统计信息",
        "operationId": "ChatController_handleChatStream",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatStreamRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "text/event-stream": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/StreamChatResponseDto"
                          },
                          "description": "流式响应数据"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "Gemini 流式聊天接口（SSE）",
        "tags": [
          "Gemini 聊天与语音"
        ]
      }
    },
    "/gemini/generate-voice": {
      "post": {
        "description": "支持多种 TTS 提供商的文字转语音服务，包括 Gemini TTS 和 MiniMax TTS",
        "operationId": "ChatController_handleGenerateVoice",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateVoiceRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "string",
                          "description": "音频文件访问地址",
                          "example": "https://example-bucket.oss-cn-shanghai.aliyuncs.com/audio/2025-12-05/output.wav"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "多提供商文本转语音接口",
        "tags": [
          "Gemini 聊天与语音"
        ]
      }
    },
    "/deepseek/chat": {
      "post": {
        "description": "接收用户消息，返回 DeepSeek 模型生成的回复，包含 token 统计信息。",
        "operationId": "DeepSeekController_handleDeepSeekChat",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpenAIChatRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/ChatResponseDto",
                          "description": "DeepSeek结果",
                          "example": {
                            "content": "你好！我是DeepSeek，很高兴为您服务！",
                            "usage": {
                              "promptTokens": 18,
                              "completionTokens": 30,
                              "totalTokens": 48
                            },
                            "model": "deepseek/deepseek-v3.2",
                            "responseTime": 800
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "DeepSeek 聊天接口",
        "tags": [
          "DeepSeek 聊天"
        ]
      }
    },
    "/deepseek/stream": {
      "post": {
        "description": "使用 DeepSeek 模型提供流式聊天服务，返回结构化响应数据，包含 token 统计信息",
        "operationId": "DeepSeekController_handleDeepSeekStream",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatStreamRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "text/event-stream": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/StreamChatResponseDto"
                          },
                          "description": "DeepSeek 流式响应数据"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "DeepSeek 流式聊天接口（SSE）",
        "tags": [
          "DeepSeek 聊天"
        ]
      }
    },
    "/gpt/chat": {
      "post": {
        "description": "接收用户消息，返回 GPT 模型生成的回复，包含 token 统计信息。",
        "operationId": "GPTController_handleGPTChat",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpenAIChatRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/ChatResponseDto",
                          "description": "GPT结果",
                          "example": {
                            "content": "你好！我是GPT，很高兴为您服务！",
                            "usage": {
                              "promptTokens": 16,
                              "completionTokens": 28,
                              "totalTokens": 44
                            },
                            "model": "pa/gt-4p",
                            "responseTime": 900
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "GPT 聊天接口",
        "tags": [
          "GPT 聊天"
        ]
      }
    },
    "/gpt/stream": {
      "post": {
        "description": "使用 GPT 模型提供流式聊天服务，返回结构化响应数据，包含 token 统计信息",
        "operationId": "GPTController_handleGPTStream",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatStreamRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "text/event-stream": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/StreamChatResponseDto"
                          },
                          "description": "GPT 流式响应数据"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "GPT 流式聊天接口（SSE）",
        "tags": [
          "GPT 聊天"
        ]
      }
    },
...
    "/vertexai-tts/generate": {
      "post": {
        "description": "使用 Google VertexAI Text-to-Speech 服务生成高质量语音，支持 Standard 和 WaveNet 语音。",
        "operationId": "VertexAITTSController_handleVertexAITTS",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VertexAiTTSRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/VertexAITSResponseDto",
                          "description": "VertexAI TTS 语音生成结果",
                          "example": {
                            "audioUrl": "https://example-bucket.oss-cn-shanghai.aliyuncs.com/audio/2025-12-04/vertexai-tts-1704312123456.wav",
                            "voiceName": "cm-CN-Wavenet-A",
                            "characterCount": 18,
                            "generationTime": 1200,
                            "sampleRate": 24000
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "VertexAI TTS 语音合成接口",
        "tags": [
          "VertexAI TTS"
        ]
      }
    },
...
  },
  "info": {
    "title": "Gemini Api 接口文档",
    "description": "基于 NestJS 实现的聊天与语音生成 API 服务",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/api"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "ResponseDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "example": 0,
            "description": "业务状态码，0 表示成功"
          },
          "msg": {
            "type": "string",
            "example": "success",
            "description": "提示信息"
          },
          "data": {
            "type": "object",
            "description": "返回数据"
          }
        },
        "required": [
          "code",
          "msg",
          "data"
        ]
      },
      "TokenUsageDto": {
        "type": "object",
        "properties": {
          "promptTokens": {
            "type": "number",
            "example": 25,
            "description": "输入提示的 token 数量"
          },
          "completionTokens": {
            "type": "number",
            "example": 150,
            "description": "生成的响应 token 数量"
          },
          "totalTokens": {
            "type": "number",
            "example": 175,
            "description": "总 token 数量"
          }
        },
        "required": [
          "promptTokens",
          "completionTokens",
          "totalTokens"
        ]
      },
      "ChatResponseDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "example": "你好！我是AI助手，很高兴为您服务！",
            "description": "AI 生成的回复内容"
          },
          "usage": {
            "description": "Token 使用统计",
            "allOf": [
              {
                "$ref": "#/components/schemas/TokenUsageDto"
              }
            ]
          },
          "model": {
            "type": "string",
            "example": "gemini-2.5-flash",
            "description": "使用的模型名称"
          },
          "responseTime": {
            "type": "number",
            "example": 1200,
            "description": "响应耗时（毫秒）"
          }
        },
        "required": [
          "content",
          "usage",
          "model",
          "responseTime"
        ]
      },
      "ChatMessageParamsDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "你好世界",
            "description": "用户输入的消息文本"
          }
        },
        "required": [
          "message"
        ]
      },
      "StreamChatResponseDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "example": "你好！我是AI助手，很高兴为您服务！",
            "description": "AI 生成的部分回复内容"
          },
          "usage": {
            "description": "当前 Token 使用统计（流式中可能为估算值）",
            "allOf": [
              {
                "$ref": "#/components/schemas/TokenUsageDto"
              }
            ]
          },
          "model": {
            "type": "string",
            "example": "gemini-2.5-flash",
            "description": "使用的模型名称"
          },
          "responseTime": {
            "type": "number",
            "example": 1200,
            "description": "当前响应耗时（毫秒）"
          },
          "isComplete": {
            "type": "boolean",
            "example": false,
            "description": "是否为最后一个响应块"
          }
        },
        "required": [
          "content",
          "usage",
          "model",
          "responseTime",
          "isComplete"
        ]
      },
      "ChatStreamRequestDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "你好世界",
            "description": "流式输出的消息文本"
          }
        },
        "required": [
          "message"
        ]
      },
      "GenerateVoiceRequestDto": {
        "type": "object",
        "properties": {
          "text": {
            "type": "string",
            "example": "你好，欢迎使用文本转语音服务！",
            "description": "要转换为语音的文本内容"
          },
          "voiceName": {
            "type": "string",
            "example": "Kore",
            "description": "语音名称，支持的语音取决于提供商",
            "enum": [
              "Kore",
              "Puck",
              "Charon",
              "Fenrir",
              "Aoede"
            ]
          },
          "outputFile": {
            "type": "string",
            "example": "out.wav",
            "description": "输出文件名（包含文件扩展名）"
          },
          "provider": {
            "type": "string",
            "description": "TTS 提供商",
            "example": "gemini",
            "enum": [
              "gemini",
              "minimax"
            ],
            "default": "gemini"
          }
        },
        "required": [
          "text",
          "provider"
        ]
      },
      "OpenAIChatRequestDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "你好，请介绍一下你自己",
            "description": "用户输入的消息文本"
          }
        },
        "required": [
          "message"
        ]
      },
...
      "VertexAITSResponseDto": {
        "type": "object",
        "properties": {
          "audioUrl": {
            "type": "string",
            "example": "https://your-bucket.oss-region.aliyuncs.com/audio/2025-12-04/output.wav",
            "description": "生成的语音文件访问地址"
          },
          "voiceName": {
            "type": "string",
            "example": "cm-CN-Wavenet-A",
            "description": "使用的语音名称"
          },
          "characterCount": {
            "type": "number",
            "example": 42,
            "description": "文本字符数"
          },
          "generationTime": {
            "type": "number",
            "example": 1500,
            "description": "生成耗时（毫秒）"
          },
          "sampleRate": {
            "type": "number",
            "example": 32000,
            "description": "音频采样率"
          }
        },
        "required": [
          "audioUrl",
          "voiceName",
          "characterCount",
          "generationTime",
          "sampleRate"
        ]
      },
      "VertexAiTTSRequestDto": {
        "type": "object",
        "properties": {
          "text": {
            "type": "string",
            "example": "你好，欢迎使用 VertexAI 语音合成服务！",
            "description": "要转换为语音的文本内容"
          },
          "voiceName": {
            "type": "string",
            "example": "cm-CN-Wavenet-A",
            "description": "语音名称，支持 Standard 和 WaveNet 语音"
          },
          "languageCode": {
            "type": "string",
            "example": "zh-CN",
            "description": "语言代码"
          },
          "speakingRate": {
            "type": "number",
            "example": 1,
            "description": "语音语速 (0.25 - 4.0)"
          },
          "pitch": {
            "type": "number",
            "example": 0,
            "description": "音调 (-20.0 - 20.0)"
          },
          "outputFile": {
            "type": "string",
            "example": "output.wav",
            "description": "输出文件名"
          }
        },
        "required": [
          "text"
        ]
      },
...
    }
  }
}